% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metropolis_hastings}
\alias{metropolis_hastings}
\title{Metropolis Hastings Sampler}
\usage{
metropolis_hastings(w, N, B)
}
\arguments{
\item{w}{[vector]: Weights}

\item{N}{[matrix]: Number of weights}

\item{B}{[vector]: Bias term}
}
\value{
a           [vector]: Ancestors
}
\description{
Metropolis Hastings Sampler
}
\examples{
w = c(0, 0)
N = 2
B = 10
CuSMC::metropolis_hastings(w, N, B)
}
