% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVN}
\alias{MVN}
\title{MultiVariateNormal Distribution}
\usage{
MVN(mu, sigma)
}
\arguments{
\item{mu}{[vector]: Mean vector.}

\item{sigma}{[matrix]: Covariant matrix.}
}
\value{
draws         [vector]: Vector of real random draws from distribution
}
\description{
MultiVariateNormal Distribution
}
\examples{
mu = c(0, 0)
sigma = matrix(c(1, 0, 0, 1), nrow = 2)
CuSMC::MVN(mu, sigma)
}
