% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVNPDF}
\alias{MVNPDF}
\title{MultiVariateNormal Probability Density Function}
\usage{
MVNPDF(x, mu, sigma)
}
\arguments{
\item{x}{[vector]: x.}

\item{mu}{[vector]: Mean vector.}

\item{sigma}{[matrix]: Covariant matrix.}
}
\value{
[numeric]: Real value
}
\description{
MultiVariateNormal Probability Density Function
}
\examples{
x = c(0, 0)
mu = c(0, 0)
sigma = matrix(c(1, 0, 0, 1), nrow = 2)
CuSMC::MVNPDF(x, mu, sigma)
}
