% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVT}
\alias{MVT}
\title{MultiVariate T Distribution}
\usage{
MVT(mu, sigma, nu)
}
\arguments{
\item{mu}{[vector]: Location vector.}

\item{sigma}{[matrix]: Dispersion matrix.}

\item{nu}{[float]: degrees of freedom.}
}
\value{
draws         [vector]: Vector of random draws from distribution
}
\description{
MultiVariate T Distribution
}
