% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MVTPDF}
\alias{MVTPDF}
\title{MultiVariate T Probability Density Function}
\usage{
MVTPDF(x, mu, sigma, nu)
}
\arguments{
\item{x}{[vector]: x.}

\item{mu}{[vector]: Location vector.}

\item{sigma}{[matrix]: Dispersion matrix.}

\item{nu}{[float]: degrees of freedom.}
}
\value{
[numeric]: Real value
}
\description{
MultiVariate T Probability Density Function
}
