% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run}
\alias{run}
\title{Run simulations}
\usage{
run(
  N,
  d,
  timeSteps,
  Y,
  m0,
  C0,
  F,
  G,
  V,
  W,
  df,
  resampler,
  distribution,
  p = 0L
)
}
\arguments{
\item{N}{[integer]: Number of particles is solution.}

\item{d}{[integer]: Number of parameters.}

\item{timeSteps}{[integer]: Total time steps.}

\item{Y}{[matrix]: Input observations.}

\item{m0}{[vector]: Initial parameters at t=0}

\item{C0}{[matrix]: Initial covariant matrix at t=0}

\item{F}{[matrix]: Covariant matrix for scaling particle samples}

\item{G}{[matrix]: Covariant matrix for scaling particle samples}

\item{V}{[matrix]: Covariant matrix for noise??}

\item{W}{[matrix]: Covariant matrix for noise??}

\item{df}{[float]:  Degrees of freedom (for MVT)}

\item{resampler}{[string]: Resampling sampler}

\item{distribution}{[string]: Distribution for sampling particles}
}
\description{
Run simulations
}
